+++ test/Main.hs
@@ -33,7 +33,7 @@ testStructureAttributes = do
   int <- parseAttribute "int" readContent
   bool <- parseAttribute "bool" readBool
   return $ TestStructure
-    { tsName  = name
+    { tsName  = nameLocalName name
     , tsInts  = [int]
     , tsBools = pure bool }
 
@@ -156,7 +156,7 @@ specParser
   :: String                     -- ^ Name of spec
   -> Document                   -- ^ Document to parse
   -> (a -> Maybe String)        -- ^ Value checker, Nothing if all ok
-  -> DomParser Identity a                -- ^ Parser itself
+  -> DomParser Identity a       -- ^ Parser itself
   -> Spec
 specParser name doc check parser = it name $ do
   result <- either throwIO return $ runDomParser doc parser
@@ -166,9 +166,9 @@ specParser name doc check parser = it name $ do
 
 specParserEq
   :: (Eq a, Show a)
-  => String            -- ^ Name of spec
-  -> Document          -- ^ Document to parse
-  -> a                 -- ^ Value parser should return
+  => String                     -- ^ Name of spec
+  -> Document                   -- ^ Document to parse
+  -> a                          -- ^ Value parser should return
   -> DomParser Identity a       -- ^ Parser itself
   -> Spec
 specParserEq name doc a parser = specParser name doc check parser
@@ -193,7 +193,7 @@ specParserFailedPath
   :: (Show a)
   => String                     -- ^ Name of spec
   -> Document
-  -> [Text]                     -- ^ Expteced path in error
+  -> DomPath                     -- ^ Expteced path in error
   -> DomParser Identity a
   -> Spec
 specParserFailedPath name doc path parser = it name $ example $ do
@@ -226,7 +226,7 @@ combinationsSpec = do
       describe "deep" $ do
         let parser = inElem "a" $ inElem "b" textFromDom
         specParserFailedPath "docDeepMultiple2" docDeepMultiple2
-          ["root", "a", "b"] parser
+          (DomPath ["root", "a", "b"]) parser
         -- should fail here because first tag "a" does not contains
         -- tag "b" which is expected by parser @inElem "b"@
 
@@ -276,7 +276,7 @@ combinationsSpec = do
         specParserEq "docDeepMultiple1" docDeepMultiple1 "content1" parser
         specParserEq "docDeepMultiple2" docDeepMultiple2 "content1" parser
       describe "fails" $ do
-        specParserFailedPath "docSimple" docSimple ["root", "a", "b"] parser
+        specParserFailedPath "docSimple" docSimple (DomPath ["root", "a", "b"]) parser
     describe "multiple elements" $ do
       let
         parser = diveElem "a" $ inElemAll "b" textFromDom
@@ -311,9 +311,9 @@ combinationsSpec = do
         inElem "a" $ do
           checkCurrentName "a"
     describe "fails" $ do
-      specParserFailedPath "root element" docSimple ["toor"] $ do
+      specParserFailedPath "root element" docSimple (DomPath ["toor"]) $ do
         checkCurrentName "toor"
-      specParserFailedPath "inner element" docSimple ["root", "b"] $ do
+      specParserFailedPath "inner element" docSimple (DomPath ["root", "b"]) $ do
         inElem "a" $ checkCurrentName "b"
 
 contentSpec :: Spec