+++ CHANGELOG.md
@@ -1,10 +1,26 @@
 # CHANGELOG
 
+## 2.0.0
+
+### Added
+
+* `NameMatcher` and `ElemMatcher` as more flexible mechanisms for
+  matching names and elements instead of just matching by name.
+* Class `FromAttribute` and some default instances
+
+### Changed
+
+* `DomTraversable` renamed to `Buildable`
+* Combinators use `NameMatcher` and `ElemMatcher`
+* Documentation improved
+
 ## 1.0.0
+
 * Added support for parsing attributes
 * Changed public interface for 'parseContent'
-* Added usefull functions
+* Added useful functions
 * Dropped unnecessary dependency from shakespeare
 
 ## 0.1.0
+
 * Public release